cloud/pipelines/aggregate/requirements.txt
faust-streaming
influxdb-client